AIPerformanceFrameowork/pdf-parser-service/requirements.txt
Flask
Werkzeug
camelot-py[cv]
pandas
openpyxl
ghostscript
